Convert ffmpeg to mythffmpeg
This will use the new mythffmpeg that is built as part of the mythtv build.
This lets us drastically simplify the command-line argument handling, as it
will be static between ffmpeg syncs.

Unfortunately, at this time, there is no MP3 audio encode support in mythffmpeg
but I will see what we can do.