Reworked ffmpeg exporters to work with mythffmpeg
Although there is doubtless more cleanup to do as a result of this, nuvexport
now uses the ffmpeg build provided by mythtv itself as part of it builds.  This
is as of MythTV/mythtv@8c6de4675d6d3ba234999611d35901953074b1d2
All of the ffmpeg exporters have been tested with this setup.

Additionally, sox is now used to downsample 6-channel audio to 2-channel before
encoding.  This should fix the reported problems of encoding from AC-3 audio.
I may refactor the dependency later so it only checks for sox if it's needed
for the particular job at hand.

Split the "MP4 (iPod)" into two exporters: MP4 and H.264, as it is confusing
to look for H.264 buried inside MP4.

Fixes #9199.  Fixes #9205.